create surve component
