+++ src/shared/components/sections/survey/clickOption.js
